requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
bcrypt==4.0.1
certifi==2023.7.22
cfgv==3.4.0
charset-normalizer==3.3.1
curlify==2.2.1
distlib==0.3.7
exceptiongroup==1.1.3
Faker==15.1.5
filelock==3.13.1
identify==2.5.32
idna==3.4
imap-tools==1.5.0
iniconfig==2.0.0
nodeenv==1.8.0
packaging==23.2
platformdirs==4.0.0
pluggy==1.3.0
postgres==4.0
pre-commit==3.5.0
psycopg2-binary==2.9.6
psycopg2-pool==1.1
PyHamcrest==2.0.4
PyJWT==2.8.0
pytest==8.0.2
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
tomli==2.0.1
urllib3==2.0.7
virtualenv==20.24.7

assertpy~=1.1
